+++ uefi/CHANGELOG.md
@@ -7,6 +7,8 @@
 - Added `proto::pci::root_bridge::PciRootBridgeIo::enumerate()`.
 - Added `proto::nvme::pass_thru::NvmePassThru::broadcast()`.
 - Added `proto::media::block::BlockIO2`.
+- Added `proto::device_path::DevicePath::to_pool()`.
+- Added `proto::device_path::DevicePathUtilities::duplicate_path()`.
 
 ## Changed
 - export all `text::{input, output}::*` types